Change all double initialization to float
This way we can squeeze a little more speed with the tradeof of presision

+++ include/kernels.cuh
@@ -19,60 +19,60 @@ inline void gpuAssert(cudaError_t code, const char *file, int line, bool abort=t
 typedef struct {
     int input_size;
     int output_size;
-    double *weights;
-    double *biases;
-    double *preactivations;
-    double *activations;
-    double *grad_wrt_w;
-    double *grad_wrt_b;
-    double *grad_wrt_input;
+    float *weights;
+    float *biases;
+    float *preactivations;
+    float *activations;
+    float *grad_wrt_w;
+    float *grad_wrt_b;
+    float *grad_wrt_input;
 } Layer;
 /**
  */
-__global__ void safe_softmax_kernel(double* input, int size, double* output, int batch_size);
+__global__ void safe_softmax_kernel(float* input, int size, float* output, int batch_size);
 
 /* Softmax kernel for SGD */
-__global__ void softmax_kernel(double* input, int size, double* output_dest);
+__global__ void softmax_kernel(float* input, int size, float* output_dest);
 
-__global__ void forward_kernel2d(double *input, double *weights, double *biases,
-                                 double *preactivations, double *activations,
+__global__ void forward_kernel2d(float *input, float *weights, float *biases,
+                                 float *preactivations, float *activations,
                                  int input_size, int output_size,
                                  int batch_size);
 
-__global__ void forward_kernel(double *d1, double *d2, double *d3, double *d4,
-                               double *d5, int i1, int i2);
+__global__ void forward_kernel(float *d1, float *d2, float *d3, float *d4,
+                               float *d5, int i1, int i2);
 
-__global__ void zero_layer_gradients(double *grad_wtr_b, double *grad_wtr_input,
+__global__ void zero_layer_gradients(float *grad_wtr_b, float *grad_wtr_input,
                                      int size);
 
-__global__ void zero_weight_gradients(double *weights, int size);
+__global__ void zero_weight_gradients(float *weights, int size);
 
 __global__ void backward_kernel_output_layer2d(
-    double *grad_wtr_b, double *grad_wtr_w, double *prev_grad_wrt_input,
-    double *previous_activation, double *weight, int input_size,
-    int output_size, int batch_size, int *target, double *probs);
+    float *grad_wtr_b, float *grad_wtr_w, float *prev_grad_wrt_input,
+    float *previous_activation, float *weight, int input_size,
+    int output_size, int batch_size, int *target, float *probs);
 
 __global__ void backward_kernel_output_layer(
-    double *grad_wrt_b, double *grad_wrt_w, double *previous_grad_wrt_input,
-    double *previous_activations, double *weights, int input_size,
-    int output_size, int target, double *probs);
+    float *grad_wrt_b, float *grad_wrt_w, float *previous_grad_wrt_input,
+    float *previous_activations, float *weights, int input_size,
+    int output_size, int target, float *probs);
 
 __global__ void backward_kernel2d(Layer *layer, Layer *previous_layer,
-                                  double *initial_input, int current_layer_idx,
+                                  float *initial_input, int current_layer_idx,
                                   int batch_size);
 
-__global__ void sparce_categorical_cross_entropy_kernel(double* probabilities, int* target_labels,
-                                                 int batch_size, int num_classes, double* loss);
+__global__ void sparce_categorical_cross_entropy_kernel(float* probabilities, int* target_labels,
+                                                 int batch_size, int num_classes, float* loss);
 
 __global__ void backward_kernel(Layer *layer, Layer *previous_layer,
-                                double *initial_inputs, int current_layer_idx);
+                                float *initial_inputs, int current_layer_idx);
 
-__global__ void update_kernel_minibatch(double* weights, double* biases, double* grad_w,
-                                        double* grad_b, int input_size, int out_size,
-                                        double learning_rate, int batch_size);
+__global__ void update_kernel_minibatch(float* weights, float* biases, float* grad_w,
+                                        float* grad_b, int input_size, int out_size,
+                                        float learning_rate, int batch_size);
 
-__global__ void update_kernel(double *weights, double *biases, double *grad_w,
-                              double *grad_b, int in_size, int out_size,
-                              double learning_rate);
+__global__ void update_kernel(float *weights, float *biases, float *grad_w,
+                              float *grad_b, int in_size, int out_size,
+                              float learning_rate);
 
 #endif // KERNELS_CUH_
